fix: Fix incorrect assumption about clickhouse query order
To understand why this is needed you gotta understand how some of this stuff works.

1. We're joining two different views here and if we want Clickhouse to join two queries we need their column order to be the exact same - and have matching types.

2. When we run `SELECT * FROM saved_query` we'll explode the star to a list of fields in our HogQL `Resolver`. The way we do that is by looking at the saved query `columns` field and listing them in order. For the old managed views we'd simply look at the `fields` field.

The problem here happens with the fact that `jsonb` columns do NOT store columns in the same order as they were inserted. It instead stores them in a query-efficient way which, to their own definition of query-efficient, implies in storing the keys in length-order (and alphabetical for same length). This means that we get a different column order when querying from a `SavedView` than we did with an inline managed view.

To solve that problem, we're now explicitly listing the columns we want in the exact same order we would expect it to be were it not scrambled by Postgres.